core/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
commonmark==0.9.1
coverage==6.3.3
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-appconf==1.0.5
django-ckeditor==6.4.0
django-filter==21.1
django-image-cropping==1.7
django-imagekit==4.1.0
django-import-export==2.8.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-modeltranslation==0.18.2
django-taggit==3.0.0
easy-thumbnails==2.8.1
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==13.11.1
inflection==0.5.1
iniconfig==1.1.1
lxml==4.8.0
Mako==1.2.0
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pilkit==2.0
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.3.0
python-dateutil==2.8.2
pywatchman==1.4.1
PyYAML==6.0
reportlab==3.6.9
rich==12.4.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
tomli==2.0.1
typing-extensions==4.2.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
